Unified tests into computation, construction, and junctiontree components
