+++ tests/test_construction.py
@@ -797,3 +797,104 @@ def test_join_trees_with_multiple_cliques_with_both_nested():
 
 
     assert_junction_trees_equal(output, merged_tree)
+
+
+def test_traversal():
+    tree = [
+        0,
+        (
+            1,
+            [
+                2,
+            ]
+        ),
+        (
+            3,
+            [
+                4,
+                (
+                    5,
+                    [
+                        6,
+                    ]
+                )
+            ]
+        )
+    ]
+
+
+    assert list(cons.bf_traverse(tree)) == [0,1,3,2,4,5,6,]
+
+    assert list(cons.df_traverse(tree)) == [0,1,2,3,4,5,6,]
+
+
+def test_get_clique_vars():
+    node_list = [
+                    [0, 2, 4],
+                    [0, 2],
+                    [0, 1, 2],
+                    [4],
+                    [3, 4],
+                    [3],
+                    [1, 2, 3]
+
+    ]
+    assert cons.get_clique_vars(node_list, 0) == [0, 2, 4]
+    assert cons.get_clique_vars(node_list, 1) == [0, 2]
+    assert cons.get_clique_vars(node_list, 2) == [0, 1, 2]
+    assert cons.get_clique_vars(node_list, 3) == [4]
+    assert cons.get_clique_vars(node_list, 4) == [3, 4]
+    assert cons.get_clique_vars(node_list, 5) == [3]
+    assert cons.get_clique_vars(node_list, 6) == [1, 2, 3]
+    assert cons.get_clique_vars(node_list, 7) == None
+
+def test_generate_potential_pairs():
+    tree = [
+        0,
+        (
+            1,
+            [
+                2,
+            ]
+        ),
+        (
+            3,
+            [
+                4,
+            ]
+        ),
+        (
+            5,
+            [
+                6,
+                (
+                    7,
+                    [
+                        8,
+                        (
+                            9,
+                            [
+                                10,
+                            ]
+                        )
+                    ]
+                )
+            ]
+
+        )
+    ]
+
+    assert cons.generate_potential_pairs(tree) == [
+                                        (0, 1),
+                                        (0, 3),
+                                        (0, 5),
+                                        (1, 2),
+                                        (3, 4),
+                                        (5, 6),
+                                        (6, 7),
+                                        (7, 8),
+                                        (8, 9),
+                                        (9, 10)
+    ]
+
+
